feat(content): Add telemetry action
