+++ content-src/components/Spotlight/Spotlight.scss
@@ -47,6 +47,7 @@
     max-height: $spotlight-title-max-height;
     overflow: hidden;
   }
+
   &:hover .spotlight-title {
     color: $link-blue;
   }
